% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureAnalysis.R
\name{MixtureAnalysis}
\alias{MixtureAnalysis}
\title{Analysis of Mixture Experiment}
\usage{
MixtureAnalysis(DF, resp = "y")
}
\arguments{
\item{DF}{Data Frame with design and Response}

\item{resp}{Character name of response column}
}
\value{
a list
}
\description{
Analysis of Mixture Experiment
}
\examples{
des <- SLDnew(3, 2, center=1, axial=1)
extension <- c(415, 484,424,400,409,342,456, 460,450, 411)
desrun1 <- data.frame(0.2+0.4*des, y=extension)
MixtureAnalysis(desrun1)
}
