% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLDnew.R
\name{SLDnew}
\alias{SLDnew}
\title{Create Simplex Lattice Design}
\usage{
SLDnew(fac, lev, center = F, axial = F)
}
\arguments{
\item{center}{logical: Is center point required; or numerical: number of center points}

\item{axial}{logical: Are axial points required}

\item{Number}{of Levels per Factor}
}
\value{
The simplex lattice design
}
\description{
Create Simplex Lattice Design
}
\examples{
SLDnew(3, 2, center=T, axial=T)
}
