% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLDnew.R
\name{SLDnew}
\alias{SLDnew}
\title{Create Simplex Lattice Design}
\usage{
SLDnew(fac, lev, center = FALSE, axial = FALSE)
}
\arguments{
\item{fac}{Number of Factors}

\item{lev}{Number of Levels per Factor}

\item{center}{logical: Is center point required; or numerical: number of center points}

\item{axial}{logical: Are axial points required}
}
\value{
The simplex lattice design
}
\description{
Create Simplex Lattice Design
}
\examples{
SLDnew(3, 2, center=FALSE, axial=FALSE)
SLDnew(3, 2, center=1, axial=1)
}
